# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MediaLibrary {
  /** Filenames in the public rooms directory */
  rooms: string[];
  /** Filenames in the public art directory */
  art: string[];
}

export interface ArtImageUpload {
  /**
     * Suggested file stem, usually slugified from the piece's name. The server sanitises it and appends a suffix if the name is taken, so it can never escape the art directory or overwrite an existing image.
     * @minLength 1
     */
  baseName: string;
  /** Data URL of the image as uploaded, at its original size */
  fullImage: string;
  /** Data URL of the tray-sized image generated from it */
  thumbnail: string;
}

export interface ArtImageFiles {
  fullImageFilename: string;
  thumbnailFilename: string;
}

export interface Room {
  id: string;
  name: string;
  /** A file in the public rooms directory. 1600x1000px, 16:10. */
  imageFilename: string;
  /**
     * Percentage of canvas height. Wall art may only be placed above it, sculptures only below it.
     * @minimum 1
     * @maximum 99
     */
  bandSplit: number;
}

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  imageFilename: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  bandSplit: number;
}

export interface RoomUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  imageFilename?: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  bandSplit?: number;
}

export type ObjectType = typeof ObjectType[keyof typeof ObjectType];


export const ObjectType = {
  wall: 'wall',
  sculpture: 'sculpture',
} as const;

export interface ArtObject {
  id: string;
  name: string;
  type: ObjectType;
  thumbnailFilename: string;
  fullImageFilename: string;
  /**
     * width / height
     * @exclusiveMinimum 0
     */
  aspectRatio: number;
  /**
     * Fraction of the room canvas width the piece occupies
     * @maximum 1
     * @exclusiveMinimum 0
     */
  defaultScale: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  minScale: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  maxScale: number;
}

export interface ArtInput {
  /** @minLength 1 */
  name: string;
  type: ObjectType;
  /** @minLength 1 */
  thumbnailFilename: string;
  /** @minLength 1 */
  fullImageFilename: string;
  /** @exclusiveMinimum 0 */
  aspectRatio: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  defaultScale: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  minScale: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  maxScale: number;
}

export interface ArtUpdate {
  /** @minLength 1 */
  name?: string;
  type?: ObjectType;
  /** @minLength 1 */
  thumbnailFilename?: string;
  /** @minLength 1 */
  fullImageFilename?: string;
  /** @exclusiveMinimum 0 */
  aspectRatio?: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  defaultScale?: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  minScale?: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  maxScale?: number;
}

export interface Placement {
  objectId: string;
  roomId: string;
  /** Percentage of canvas width, 0-100 */
  x: number;
  /** Percentage of canvas height, 0-100 */
  y: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  scale: number;
}

export type PlacementSet = Placement[];