# lib/api-zod/src/generated/types/artImageUpload.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArtImageUpload {
  /**
     * Suggested file stem, usually slugified from the piece's name. The server sanitises it and appends a suffix if the name is taken, so it can never escape the art directory or overwrite an existing image.
     * @minLength 1
     */
  baseName: string;
  /** Data URL of the image as uploaded, at its original size */
  fullImage: string;
  /** Data URL of the tray-sized image generated from it */
  thumbnail: string;
}