# lib/api-zod/src/generated/types/artObject.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObjectType } from './objectType';

export interface ArtObject {
  id: string;
  name: string;
  type: ObjectType;
  thumbnailFilename: string;
  fullImageFilename: string;
  /**
     * width / height
     * @exclusiveMinimum 0
     */
  aspectRatio: number;
  /**
     * Fraction of the room canvas width the piece occupies
     * @maximum 1
     * @exclusiveMinimum 0
     */
  defaultScale: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  minScale: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  maxScale: number;
}