# lib/api-zod/src/generated/types/artUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObjectType } from './objectType';

export interface ArtUpdate {
  /** @minLength 1 */
  name?: string;
  type?: ObjectType;
  /** @minLength 1 */
  thumbnailFilename?: string;
  /** @minLength 1 */
  fullImageFilename?: string;
  /** @exclusiveMinimum 0 */
  aspectRatio?: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  defaultScale?: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  minScale?: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  maxScale?: number;
}