# lib/api-zod/src/generated/types/mediaLibrary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MediaLibrary {
  /** Filenames in the public rooms directory */
  rooms: string[];
  /** Filenames in the public art directory */
  art: string[];
}