# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: string;
  name: string;
  /** A file in the public rooms directory. 1600x1000px, 16:10. */
  imageFilename: string;
  /**
     * Percentage of canvas height. Wall art may only be placed above it, sculptures only below it.
     * @minimum 1
     * @maximum 99
     */
  bandSplit: number;
}