# lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  imageFilename: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  bandSplit: number;
}