# lib/api-zod/src/generated/types/roomUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoomUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  imageFilename?: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  bandSplit?: number;
}